Restore a private ValueInEmptyWindows to DataflowRunner for special hacks
